% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.test_mediation}
\alias{coef.test_mediation}
\alias{coef.boot_test_mediation}
\alias{coef.fit_mediation}
\title{Coefficients in (robust) mediation analysis}
\usage{
\method{coef}{test_mediation}(object, parm = NULL, ...)

\method{coef}{boot_test_mediation}(object, parm = NULL, type = c("boot",
  "data"), ...)

\method{coef}{fit_mediation}(object, parm = NULL, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"\link{test_mediation}"}
containing results from (robust) mediation analysis, or an object inheriting
from class \code{"\link{fit_mediation}"} containing a (robust) mediation
model fit.}

\item{parm}{an integer, character or logical vector specifying the
coefficients to be extracted, or \code{NULL} to extract all coefficients.}

\item{\dots}{additional arguments are currently ignored.}

\item{type}{a character string specifying whether to extract the means
of the bootstrap distribution (\code{"boot"}; the default), or the
coefficient estimates based on the full data set (\code{"data"}).}
}
\value{
A numeric vetor containing the requested coefficients.
}
\description{
Extract coefficients from models computed in (robust) mediation analysis.
}
\seealso{
\code{\link{test_mediation}}, \code{\link{fit_mediation}},
\code{\link[=confint.test_mediation]{confint}}
}
\author{
Andreas Alfons
}
\keyword{utilities}
