\name{mediate-package}
\alias{mediate-package}
\docType{package}
\title{
(Robust) mediation analysis
}
\description{
Perform mediation analysis via a (fast and robust) bootstrap test.
}
\details{
\tabular{ll}{
Package: \tab mediate\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-07-11\cr
Depends: \tab boot (>= 1.3-1), ggplot2 (>= 0.9.3), robustbase (>= 0.9-8)\cr
License: \tab GPL (>= 2)\cr
LazyData: \tab yes\cr
}

Index:
\preformatted{
coef.bootMA             Coefficients in (robust) mediation analysis
confint.bootMA          Confidence intervals for (robust) mediation
                        analysis
covHuber                Huber-type M-estimator of location and scatter
covHuber.control        Tuning parameters for Huber-type M-estimation
                        of location and scatter
fortify.bootMA          Convert (robust) mediation analysis results
                        into a data frame for plotting
mediate                 (Robust) mediation analysis
mediate-package         (Robust) mediation analysis
mediatePlot             Plot (robust) mediation analysis results
retest                  Retest for mediation
}
}
\author{
Andreas Alfons [aut, cre]

Maintainer: Andreas Alfons <alfons@ese.eur.nl>
}
\keyword{package}
