\name{mediate}
\alias{mediate}
\alias{print.bootMA}
\alias{print.sobelMA}
\alias{summary.bootMA}
\alias{summary.sobelMA}
\title{(Robust) mediation analysis}
\usage{
  mediate(x, y, m, method = c("boot", "sobel"),
    alternative = c("twosided", "less", "greater"),
    R = 5000, level = 0.95, type = c("bca", "perc"),
    robust = TRUE, transform = FALSE, control, ...)
}
\arguments{
  \item{x}{a numeric vector containing the independent
  variable.}

  \item{y}{a numeric vector containing the dependent
  variable.}

  \item{m}{a numeric vector containing the proposed
  mediator variable.}

  \item{method}{a character string specifying the test to
  be performed for the indirect effect.  Possible values
  are \code{"boot"} (the default) for the bootstrap, or
  \code{"sobel"} for Sobel's test.}

  \item{alternative}{a character string specifying the
  alternative hypothesis in the test for the indirect
  effect.  Possible values are \code{"twosided"} (the
  default), \code{"less"} or \code{"greater"}.}

  \item{R}{an integer giving the number of bootstrap
  replicates.  The default is to use 5000 bootstrap
  replicates.}

  \item{level}{numeric; the confidence level of the
  confidence interval in the bootstrap test.  The default
  is to compute a 95\% confidence interval.}

  \item{type}{a character string specifying the type of
  confidence interval to be computed in the bootstrap test.
  Possible values are \code{"bca"} (the default) for the
  bias-corrected and accelerated bootstrap, or
  \code{"perc"} for the percentile bootstrap.}

  \item{robust}{a logical indicating whether to perform a
  robust test (defaults to \code{TRUE}).}

  \item{transform}{logical; if \code{robust} is
  \code{TRUE}, this indicates whether to clean the data
  with a Huber-type transformation (defaults to
  \code{FALSE}).}

  \item{control}{if \code{robust} is \code{TRUE} and
  \code{transform} is \code{FALSE}, a list of tuning
  parameters for \code{\link[robustbase]{lmrob}} as
  generated by \code{\link[robustbase]{lmrob.control}}.  If
  both \code{robust} and \code{transform} are \code{TRUE},
  a list of tuning parameters for \code{\link{covHuber}} as
  generated by \code{\link{covHuber.control}}.}

  \item{\dots}{additional arguments to be passed to
  \code{\link[boot]{boot}} in case of a bootstrap test.}
}
\value{
  An object of class \code{"bootMA"} (if \code{method} is
  \code{"boot"}) or \code{"sobelMA"} (if \code{method} is
  \code{"sobel"}) with the following components:

  \item{ab}{numeric; the point estimate of the indirect
  effect.}

  \item{ci}{a numeric vector of length two containing the
  bootstrap confidence interval for the indirect effect
  (only \code{"bootMA"}).}

  \item{reps}{an object of class \code{"\link[boot]{boot}"}
  containing the bootstrap replicates of the indirect
  effect (only \code{"bootMA"}).}

  \item{se}{numeric; the standard error of the indirect
  effect according to Sobel's formula (only
  \code{"sobelMA"}).}

  \item{statistic}{numeric; the test statistic for Sobel's
  test (only \code{"sobelMA"}).}

  \item{pValue}{numeric; the p-Value from Sobel's test
  (only \code{"sobelMA"}).}

  \item{alternative}{a character string specifying the
  alternative hypothesis in the test for the indirect
  effect.}

  \item{R}{an integer giving the number of bootstrap
  replicates (only \code{"bootMA"}).}

  \item{level}{numeric; the confidence level of the
  bootstrap confidence interval (only \code{"bootMA"}).}

  \item{type}{a character string specifying the type of
  bootstrap confidence interval (only \code{"bootMA"}).}

  \item{robust}{a logical indicating whether a robust test
  was performed.}

  \item{transform}{a logical indicating whether the data
  were cleaned with a Huber-type transformation.}

  \item{fitYX}{an object of class
  \code{"\link[robustbase]{lmrob}"} or
  \code{"\link[stats]{lm}"} containing the estimation
  results from the regression of the dependent variable on
  the independent variable.}

  \item{fitMX}{an object of class
  \code{"\link[robustbase]{lmrob}"} or
  \code{"\link[stats]{lm}"} containing the estimation
  results from the regression of the proposed mediator
  variable on the independent variable.}

  \item{fitYMX}{an object of class
  \code{"\link[robustbase]{lmrob}"} or
  \code{"\link[stats]{lm}"} containing the estimation
  results from the regression of the dependent variable on
  the proposed mediator and independent variables.}

  \item{data}{a data frame containing the indipendent,
  dependent and proposed mediator variables.}
}
\description{
  Perform (robust) mediation analysis via a (fast and
  robust) bootstrap test or Sobel's test.
}
\details{
  If \code{robust} is \code{TRUE} and \code{transform} is
  \code{FALSE} (the defaults), the tests are based on
  robust regressions with \code{\link[robustbase]{lmrob}}.
  The bootstrap test is thereby performed via the fast and
  robust bootstrap.

  If both \code{robust} and \code{transform} are
  \code{TRUE}, the data are first cleaned with a Huber-type
  transformation.  Then standard tests are performed with
  the cleaned data.  Note that this approach is less robust
  than the one described above.  Furthermore, the bootstrap
  does not account for the variability from cleaning the
  data.
}
\examples{
data("superbowl")
with(superbowl, mediate(frequency, liking, clutter))
}
\author{
  Andreas Alfons
}
\references{
  Preacher, K.J. and Hayes, A.F. (2004) SPSS and SAS
  procedures for estimating indirect effects in simple
  mediation models. \emph{Behavior Research Methods,
  Instruments, & Computers}, \bold{36}(4), 717--731.

  Salibian-Barrera, M. and Zamar, R. (2002) Bootstrapping
  robust estimates of regression. \emph{The Annals of
  Statistics}, \bold{30}(2), 556--582.

  Sobel, M.E. (1982) Asymptotic confidence intervals for
  indirect effects in structural equation models.
  \emph{Sociological Methodology}, \bold{13}, 290--312.

  Zu, J. and Yuan, K.-H. (2010) Local influence and robust
  procedures for mediation analysis. \emph{Multivariate
  Behavioral Research}, \bold{45}(1), 1--44.
}
\seealso{
  \code{\link[=coef.bootMA]{coef}},
  \code{\link[=confint.bootMA]{confint}},
  \code{\link[=fortify.bootMA]{fortify}} and
  \code{\link[=mediatePlot]{plot}} methods

  \code{\link[boot]{boot}},
  \code{\link[robustbase]{lmrob}}, \code{\link[stats]{lm}},
  \code{\link{covHuber}}
}
\keyword{multivariate}

