\name{robmed-package}
\alias{robmed-package}
\docType{package}
\title{
\packageTitle{robmed}
}
\description{
\packageDescription{robmed}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{robmed}
\packageIndices{robmed}
}
\author{
\packageAuthor{robmed}

Maintainer: \packageMaintainer{robmed}
}
\keyword{package}
