% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_weight_plot.R
\name{setup_weight_plot}
\alias{setup_weight_plot}
\alias{setup_weight_plot.test_mediation}
\alias{setup_weight_plot.reg_fit_mediation}
\alias{setup_weight_plot.cov_fit_mediation}
\title{Set up a diagnostic plot of regression weights}
\usage{
setup_weight_plot(object, ...)

\method{setup_weight_plot}{test_mediation}(object, ...)

\method{setup_weight_plot}{reg_fit_mediation}(object, outcome = NULL, npoints = 1000, ...)

\method{setup_weight_plot}{cov_fit_mediation}(object, ...)
}
\description{
Extract the relevant information for a diagnostic plot of the regression
weights from (robust) mediation analysis.  This plot allows to easily detect
deviations from normality assumptions such as skewness or heavy tails.
}
\details{
This function is used internally by \code{\link{weight_plot}()}.  It may
also be useful for users who want to produce a similar plot, but who want
more control over what information to display or how to display that
information.
}
\examples{
data("BSG2014")

# run fast and robust bootstrap test
robust_boot <- test_mediation(BSG2014,
                              x = "ValueDiversity",
                              y = "TeamCommitment",
                              m = "TaskConflict",
                              robust = TRUE)

# set up information for plot
setup <- setup_weight_plot(robust_boot)

}
\seealso{
\code{\link{fit_mediation}()}, \code{\link{test_mediation}()},
\code{\link{weight_plot}()}
}
\author{
Andreas Alfons
}
\keyword{hplot}
