% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_plot.R
\name{weight_plot}
\alias{weight_plot}
\alias{weight_plot.default}
\alias{weight_plot.setup_weight_plot}
\title{Diagnostic plot of robust regression weights}
\usage{
weight_plot(object, ...)

\method{weight_plot}{default}(object, outcome = NULL, npoints = 1000, ...)

\method{weight_plot}{setup_weight_plot}(object, ...)
}
\value{
An object of class \code{"\link[ggplot2]{ggplot}"}.
}
\description{
Produce a diagnostic plot of the regression weights from robust mediation
analysis.  This plot allows to easily detect deviations from normality
assumptions such as skewness or heavy tails.
}
\examples{
data("BSG2014")

# run fast and robust bootstrap test
robust_boot <- test_mediation(BSG2014,
                              x = "ValueDiversity",
                              y = "TeamCommitment",
                              m = "TaskConflict",
                              robust = TRUE)

# create plot for robust bootstrap test
weight_plot(robust_boot)

}
\seealso{
\code{\link{fit_mediation}()}, \code{\link{test_mediation}()},
\code{\link{setup_weight_plot}()}

\code{\link{ci_plot}()}, \code{\link{density_plot}()},
\code{\link{ellipse_plot}()}, \code{\link[=plot-methods]{plot}()}
}
\author{
Andreas Alfons
}
\keyword{hplot}
