% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights.cov_Huber}
\alias{weights.cov_Huber}
\title{Robustness weights of Huber M-estimation of location and scatter}
\usage{
\method{weights}{cov_Huber}(object, type = c("consistent", "relative"), ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"\link{cov_Huber}"}
containing Huber M-estimates of location and scatter.}

\item{type}{a character string specifying the type of robustness weights to
be extracted.  Possible values are \code{"consistent"} and
\code{"relative"}.  The former can be used for a robust transformation of
the data such that the covariance matrix of the transformed data is Fisher
consistent.  Observations that are not downweighted in general receive a
weight larger than 1.  The latter are useful for interpretation, as
observations that are not downweighted receive a relative weight of 1.}

\item{\dots}{additional arguments are currently ignored.}
}
\value{
A numeric vetor containing the requested robustness weights.
}
\description{
Extract (relative) robustness weights of a Huber M-estimate of location and
scatter.
}
\examples{
data("BSG2014")

# define variables
x <- "ValueDiversity"
y <- "TeamCommitment"
m <- "TaskConflict"

# compute Huber M-estimator
S <- cov_Huber(BSG2014[, c(x, y, m)])
weights(S, type = "relative")

}
\references{
Zu, J. and Yuan, K.-H. (2010) Local influence and robust procedures for
mediation analysis. \emph{Multivariate Behavioral Research}, \bold{45}(1),
1--44.
}
\seealso{
\code{\link{cov_Huber}}
}
\author{
Andreas Alfons
}
\keyword{utilities}
