\name{mediate-package}
\alias{mediate-package}
\alias{mediate}
\docType{package}
\title{
(Robust) mediation analysis
}
\description{
Perform mediation analysis via a (fast and robust) bootstrap test.
}
\details{
\tabular{ll}{
Package: \tab mediate\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2013-07-11\cr
Depends: \tab ggplot2 (>= 0.9.3), robustbase (>= 0.9-8)\cr
Imports: \tab boot (>= 1.3-1)\cr
License: \tab GPL (>= 2)\cr
LazyData: \tab yes\cr
LazyLoad: \tab yes\cr
}

Index:
\preformatted{
coef.testMediation      Coefficients in (robust) mediation analysis
confint.testMediation   Confidence intervals for (robust) mediation
                        analysis
covHuber                Huber M-estimator of location and scatter
covHuber.control        Tuning parameters for Huber M-estimation of
                        location and scatter
covML                   Maximum likelihood estimator of mean vector and
                        covariance matrix
fitMediation            (Robustly) fit a mediation model
fortify.testMediation   Convert (robust) mediation analysis results
                        into a data frame for plotting
mediate-package         (Robust) mediation analysis
plotMediation           Plot (robust) mediation analysis results
retest                  Retest for mediation
testMediation           (Robust) mediation analysis
weights.covHuber        Robustness weights of Huber M-estimation of
                        location and scatter
}
}
\author{
Andreas Alfons [aut, cre]

Maintainer: Andreas Alfons <alfons@ese.eur.nl>
}
\keyword{package}
