% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robmed-deprecated.R
\name{robmed-deprecated}
\alias{robmed-deprecated}
\alias{fitMediation}
\alias{regControl}
\alias{covControl}
\alias{covHuber}
\alias{covML}
\alias{testMediation}
\alias{testMediation.default}
\alias{testMediation.fit_mediation}
\alias{plotMediation}
\title{Deprecated functions in package \pkg{robmed}}
\usage{
fitMediation(x, y, m, covariates = NULL, data, ...)

regControl(efficiency = 0.85, maxIterations = 200, tol = 1e-07,
  seed = NULL)

covControl(prob = 0.95, maxIterations = 200, tol = 1e-07)

covHuber(...)

covML(...)

testMediation(x, ...)

\method{testMediation}{default}(x, y, m, covariates = NULL, data, ...)

\method{testMediation}{fit_mediation}(x, ...)

plotMediation(...)
}
\arguments{
\item{x}{either a numeric vector containing the independent variable, or
(if \code{data} is supplied) a character string, an integer or a logical
vector specifying the corresponding column of \code{data}.}

\item{y}{either a numeric vector containing the dependent variable, or
(if \code{data} is supplied) a character string, an integer or a logical
vector specifying the corresponding column of \code{data}.}

\item{m}{either a numeric vector or data frame containing the proposed
mediator variables, or (if \code{data} is supplied) a character, integer or
logical vector specifying the corresponding columns of \code{data}.}

\item{covariates}{optional; either a numeric vector or data frame
containing additional covariates to be used as control variables, or (if
\code{data} is supplied) a character, integer or logical vector specifying
the corresponding columns of \code{data}.}

\item{data}{an optional \code{data.frame} containing the variables.}

\item{\dots}{arguments to be passed down}

\item{efficiency}{a numeric value giving the desired efficiency (defaults
to 0.85 for 85\% efficiency).}

\item{maxIterations}{an integer giving the maximum number of iterations in
iterative algorithms.}

\item{tol}{a small positive numeric value to be used to determine
convergence in various parts of the algorithm.}

\item{seed}{optional initial seed for the random number generator (see
\code{\link{.Random.seed}}).}

\item{prob}{numeric; probability for the quantile of the
\eqn{\chi^{2}}{chi-squared} distribution to be used as cutoff point in the
Huber weight function (defaults to 0.95).}
}
\description{
These functions are provided for compatibility with older versions only, and
may be defunct as soon as the next release.
}
\seealso{
\code{\link[base]{Deprecated}}

\code{\link{fit_mediation}}, \code{\link{test_mediation}},
\code{\link{plot_mediation}}

\code{\link{reg_control}}, \code{\link{cov_control}}

\code{\link{cov_Huber}}, \code{\link{cov_ML}}
}
\author{
Andreas Alfons
}
\keyword{multivariate}
