\name{robmed-package}
\alias{robmed-package}
\docType{package}
\title{
\packageTitle{robmed}
}
\description{
\packageDescription{robmed}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{robmed}
\packageIndices{robmed}
}
\examples{
## compare bootstrap methods for mediation analysis on simulated data
\dontrun{
run_shiny_app()
}

## reproduce empirical examples
\dontrun{
demo("case1")
demo("case2")
demo("case3")
}
}
\author{
\packageAuthor{robmed}

Maintainer: \packageMaintainer{robmed}
}
\keyword{package}
