% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robmed-package.R
\docType{package}
\name{robmed-package}
\alias{robmed-package}
\title{\packageTitle{robmed}}
\description{
\packageDescription{robmed}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{robmed}
\packageIndices{robmed}
}
\references{
Alfons, A., Ates, N.Y. and Groenen, P.J.F. (2022a) A Robust Bootstrap Test
for Mediation Analysis.  \emph{Organizational Research Methods},
\bold{25}(3), 591--617.  doi:10.1177/1094428121999096.

Alfons, A., Ates, N.Y. and Groenen, P.J.F. (2022b) Robust Mediation Analysis:
The \R Package \pkg{robmed}.  \emph{Journal of Statistical Software},
\bold{103}(13), 1--45.  doi:10.18637/jss.v103.i13.

Alfons, A. and Schley, D.R. (2025) \emph{Robust Mediation Analysis: What We
Talk About When We Talk About Robustness}.  PsyArXiV,
doi:10.31234/osf.io/2hqdy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aalfons/robmed}
  \item Report bugs at \url{https://github.com/aalfons/robmed/issues}
}

}
\author{
\packageAuthor{robmed}

Maintainer: \packageMaintainer{robmed}
}
\keyword{package}
