\name{robmed-package}
\alias{robmed-package}
\docType{package}
\title{
\packageTitle{robmed}
}
\description{
\packageDescription{robmed}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{robmed}
\packageIndices{robmed}
}
\examples{
## compare bootstrap methods for mediation analysis on simulated data
\dontrun{
run_shiny_app()
}

## reproduce empirical examples
\dontrun{
demo("case1")
demo("case2")
demo("case3")
}
}
\references{
Alfons, A., Ates, N.Y. and Groenen, P.J.F. (2021) A robust bootstrap test
for mediation analysis.  \emph{Organizational Research Methods},
\doi{10.1177/1094428121999096}.
}
\author{
\packageAuthor{robmed}

Maintainer: \packageMaintainer{robmed}
}
\keyword{package}
