% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mediation.R
\name{sim_mediation}
\alias{sim_mediation}
\alias{sim_mediation.fit_mediation}
\alias{sim_mediation.test_mediation}
\alias{rmediation}
\title{Generate data from a fitted mediation model}
\usage{
sim_mediation(object, n, ...)

\method{sim_mediation}{fit_mediation}(
  object,
  n = NULL,
  explanatory = c("sim", "boot"),
  errors = c("sim", "boot"),
  num_discrete = 10,
  ...
)

\method{sim_mediation}{test_mediation}(object, n = NULL, ...)

rmediation(n, object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"\link{fit_mediation}"}
or \code{"\link{test_mediation}"} containing results from (robust) mediation
analysis.}

\item{n}{an integer giving the number of observations to be generated.  If
\code{NULL} (the default), the number of observations is taken from the data
set used in the fitted mediation model from \code{object}.}

\item{\dots}{additional arguments to be passed down.}

\item{explanatory}{a character string specifying how to generate the
explanatory variables (i.e., the independent variables and additional
covariates).  Possible values are \code{"sim"} to draw each explanatory
variable independently from a certain distribution (the default), or
\code{"boot"} to bootstrap the explanatory variables from the observed data
(i.e., random sampling with replacement).  See \sQuote{Details} for more
information on how the data are generated.}

\item{errors}{a character string specifying how to generate the error terms
in the linear models for the mediators and the dependent variable.  Possible
values are \code{"sim"} to draw the error terms independently from the
respective fitted model distribution (the default), or \code{"boot"} to
bootstrap the error terms from the observed residuals in the respective
fitted model (i.e., random sampling with replacement).  See \sQuote{Details}
for more information on how the data are generated.}

\item{num_discrete}{integer; if the explanatory variables are drawn
from distributions (\code{explanatory} = "sim"), variables that take
\code{num_discrete} or fewer values are considered discrete (the default is
10).  In that case, the corresponding variables are drawn from multinomial
distributions with the relative frequencies from the observed data.  This is
only relevant if the mediation model was fitted via regressions and ignored
if the mediation model was fitted via the covariance matrix, as the latter
method assumes multivariate normality.}
}
\value{
A data frame with \code{n} observations containing simulated data
for the variables of the fitted mediation model.
}
\description{
Generate data from a fitted mediation model, using the obtained coefficient
estimates as the true model coefficients for data generation.
}
\examples{
data("BSG2014")

# fit a mediation model
fit <- fit_mediation(BSG2014,
                     x = "ValueDiversity",
                     y = "TeamCommitment",
                     m = "TaskConflict")

# simulate data from the fitted mediation model
simulated_data <- sim_mediation(fit, n = 100)
head(simulated_data)

}
\seealso{
\code{\link{fit_mediation}()}, \code{\link{test_mediation}()}
}
\author{
Andreas Alfons
}
